Refactor AnalyzeInvoke()
